+++ src/CardColNet.cpp
@@ -8,7 +8,7 @@
 //REVISION    : $Revision$
 //AUTHOR      : Markus Schwab
 //CREATED     : 7.7.2005
-//COPYRIGHT   : Copyright (C) 2005, 2006
+//COPYRIGHT   : Copyright (C) 2005 - 2007
 
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -141,7 +141,7 @@ void* CardgameCollection::waitForMessages (void* thread) {
    TRACE1 ("CardgameCollection::waitForMessage (void*)");
    Check2 (cmgr.getMode () != YGP::ConnectionMgr::NONE);
 
-   int iPlayer ((int)((YGP::Thread*)thread)->getArgs ());
+   long iPlayer ((long)((YGP::Thread*)thread)->getArgs ());
    Check2 ((cmgr.getMode () == YGP::ConnectionMgr::CLIENT)
            ? (iPlayer == -1) : (iPlayer < (int)cmgr.getClients ().size ()));
